requirements.txt
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
openai==1.3.0
python-dotenv==1.0.0
